Allow compile-time detection of crypto on AArch64
The existing way depends on the libc crate, only works on Linux, and
does a getauxval() call each 64 bytes of input.

This commit adds support for doing this detection at compile-time, if
the target_feature crypto is enabled we can use the same code-path as
amd64 and reuse the asm feature.

A new compile_error!() has been added helping users figure out the best
way to build this crate.