Use new `*Dirty` traits from the `digest` crate
Updates all of the hash implementations in this repo to use the new
`*Dirty` traits which support blanket impls of the original traits,
which either consume the hasher instance or can be reset.

This will also provide a marginal efficiency boost, at least until
placement return lands (which, as it were, may be soon).