Update workspace feature resolver version to 2
Version 1 of feature resolver always causes features unification of
build and dev dependencies, which is unwanted in some circumstances.
For example, the `dev` feature of the `digest` always depends on
`alloc` which requires to define a global allocator in static libs
even though that is never used by library code.