sha3: Add Drop & ZeroizeOnDrop for Sha3State
Not zeroizing the Sha3State allows to recover any squeezed output. This
is because the `keccak` permutations can be inversed. Hence, access to
the complete state allows to perform this operation.

While this is security-relevant, including it would significantly
increase the MSRV. Therefore, it is gated behind the `zeroize` feature.